+++ BankRestApi/Data/Repositories/IAccountsRepository.cs
@@ -8,9 +8,9 @@ namespace BankRestApi.Data.Repositories
 {
     public interface IAccountsRepository
     {
-        public double? getBalance(String accountNumber);
+        public decimal? getBalance(String accountNumber);
 
-        public void updateBalance(String accountNumber, double amount); 
+        public void updateBalance(String accountNumber, decimal amount); 
         
         
     }
