+++ BankRestApi/Data/Repositories/IStatementsRepository.cs
@@ -10,6 +10,6 @@ namespace BankRestApi.Data.Repositories
     {
         public IEnumerable<StatementEntry> get(String accountNumber);
 
-        public void save(String accountNumber, DateTime date, String description, double balanceVariation, double balance);
+        public void save(String accountNumber, DateTime date, String description, decimal balanceVariation, decimal balance);
     }
 }
